+++ app/src/main/java/com/example/firebasetest/BladeAttack.java
@@ -3,7 +3,7 @@ package com.example.firebasetest;
 import android.graphics.Bitmap;
 
 public class BladeAttack extends Projectile{
-    public BladeAttack(String name, int ID, Character creator, float power, float xLocation, float yLocation, float width, float height, String effect, double direction)
+    public BladeAttack(String name, int ID, Character creator, double power, float xLocation, float yLocation, float width, float height, String effect, double direction)
     {
         super(name, ID, creator, power, 0, 0, xLocation, yLocation, width, height, direction, effect);
         this.isTimed = true;
