+++ app/src/main/java/com/example/firebasetest/LightLine.java
@@ -3,7 +3,7 @@ package com.example.firebasetest;
 import android.graphics.Bitmap;
 
 public class LightLine extends Projectile{
-    public LightLine(int ID, Character creator, float power, float xLocation, float yLocation, float width, float height, String effect, double direction)
+    public LightLine(int ID, Character creator, double power, float xLocation, float yLocation, float width, float height, String effect, double direction)
     {
         super("lightLine", ID, creator, power, 0, 0, xLocation, yLocation, width, height, direction, effect);
         this.TTD = System.currentTimeMillis() + 500L;//currently half a second. maybe a full? maybe 0.75?
@@ -13,7 +13,7 @@ public class LightLine extends Projectile{
         {
             this.TTD += 4500L;//it should hit a specific num of times per second / per 5 idk
             this.oneTimeHit = false;
-            this.power /= 30;//same dps, hits once per frame
+            this.power /= 30d;//same dps, hits once per frame
         }
     }
 }
